refactor: extract anthropic provider module
